refactor(runner): move sage-runner out of substrate/ — closes #922
`sage-runner.ts` is the `engine: sage` review runner (it shells the sage
lens-CLI), NOT an M6 HarnessId substrate. cortex#920's review flagged that
leaving it under `src/runner/substrate/` preserves the very engine/harness
conflation that PR split.

- Moved `src/runner/substrate/sage-runner.ts` → `src/runner/sage-runner.ts`
  (alongside the CC-path `review-pipeline.ts`) + its test to
  `src/runner/__tests__/`. The now-empty `substrate/` dir is removed.
- Relative imports rebased; external importers (`cortex.ts`,
  `cortex.review-consumer.e2e.test.ts`) updated.
- Refreshed the stale docblock (was "cortex#331 Phase 1 / pi-dev substrate /
  hardcode commented") to current reality: engine: sage routing, model field,
  cortex#888 block parsing. Pure move + comments — no behavior change.

616 pass across runner/boot/e2e; tsc + lint clean.
